raspi/requirements.txt
numpy
smbus2
mido
rtmidi
tabulate
pyyaml
python-rtmidi
click
tqdm
pandas
matplotlib
scikit-learn
tqdm
fastapi